+++ src/Servicio/app/models/Bio_inspired.py
@@ -6,13 +6,14 @@ from models.Cell import Cell
 from models.Member import Member
 from sqlalchemy.types import Float
 
+from sqlalchemy import Integer, Enum, BigInteger
 
 
 
 class Bio_inspired(Base):
     __tablename__ = 'Bio_inspired'
     cell_id=Column(Integer, ForeignKey(Cell.id, ondelete="CASCADE"), primary_key=True,default=True)
-    member_id=Column(Integer,ForeignKey(Member.id,ondelete="CASCADE"),primary_key=True)
+    member_id=Column(BigInteger,ForeignKey(Member.id,ondelete="CASCADE"),primary_key=True)
     threshold=Column(Float)
