+++ src/Servicio/app/models/Member.py
@@ -1,13 +1,13 @@
 from sqlalchemy import Integer, String, Column, Boolean, ForeignKey,Boolean,DateTime
 from sqlalchemy.orm import relationship
 from db.base_class import Base
-from sqlalchemy import Integer, Enum
+from sqlalchemy import Integer, Enum, BigInteger
 
 
 class Member(Base):
     __tablename__='Member'
     mail=Column(String, nullable=False)
-    id=Column(Integer, primary_key=True, index=True, unique=True,  nullable=False)
+    id=Column(BigInteger, primary_key=True, index=True, unique=True,  nullable=False)
     name=Column(String,nullable=True)
     age=Column(Integer,nullable=True)
     surname=Column(String,nullable=True)
